+++ CHANGELOG.md
@@ -42,6 +42,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
  - Fixed logic error in calculation of offset (see jgm/commonmark.js@94053a8)
  - Fixed bug where `DocParser` checked the wrong method to determine remainder handling behavior
  - Fixed bug where `HorizontalRuleParser` failed to advance the cursor beyond the parsed horizontal rule characters
+ - Fixed `DocParser` not ignoring the final newline of the input (like the reference parser does)
 
 ### Removed
  - Removed `Block\Element\AbstractInlineContainer`